requirements.txt
azure-functions
requests
beautifulsoup4
azure-storage-blob